#38 navigation refactor package
